+++ src/Request.php
@@ -683,29 +683,41 @@ class Request
     }
 
     /**
-     * Use this method to send text messages. On success, the sent Message is returned
+     * Use this method to send text messages. On success, the last sent Message is returned
+     *
+     * All message responses are saved in `$extras['responses']`.
      *
      * @link https://core.telegram.org/bots/api#sendmessage
      *
-     * @param array $data
+     * @todo Splitting formatted text may break the message.
+     *
+     * @param array      $data
+     * @param array|null $extras
      *
      * @return ServerResponse
      * @throws TelegramException
      */
-    public static function sendMessage(array $data): ServerResponse
+    public static function sendMessage(array $data, ?array &$extras = []): ServerResponse
     {
-        $text = $data['text'];
+        $text       = $data['text'];
+        $max_length = 4096;
+
+        $extras    = (array) $extras;
+        $responses = [];
 
         do {
-            //Chop off and send the first message
-            $data['text'] = mb_substr($text, 0, 4096);
-            $response     = self::send('sendMessage', $data);
+            // Chop off and send the first message.
+            $data['text'] = mb_substr($text, 0, $max_length);
+            $responses[]  = self::send('sendMessage', $data);
 
-            //Prepare the next message
-            $text = mb_substr($text, 4096);
-        } while (mb_strlen($text, 'UTF-8') > 0);
+            // Prepare the next message.
+            $text = mb_substr($text, $max_length);
+        } while ($text !== '');
 
-        return $response;
+        // Add all response objects to referenced variable.
+        $extras['responses'] = $responses;
+
+        return end($responses);
     }
 
     /**
@@ -717,7 +729,7 @@ class Request
      * @return ServerResponse
      * @throws TelegramException
      */
-    public static function __callStatic(string $action, array $data)
+    public static function __callStatic(string $action, array $data): ServerResponse
     {
         // Only argument should be the data array, ignore any others.
         return static::send($action, reset($data) ?: []);